Refactored so everything is using decimal, conversions were becoming a hassle.

+++ Application/Coin.cs
@@ -5,11 +5,11 @@ namespace VendingKataTake2
 	//The size and weight of the coin are the same as it's value for ease of use.
 	public class Coin
 	{
-		public int Weight { get; private set; }
+		public decimal Weight { get; private set; }
 
-		public int Size { get; private set; }
+		public decimal Size { get; private set; }
 
-		public Coin (int size, int weight)
+		public Coin (decimal size, decimal weight)
 		{
 			Weight = weight;
 			Size = size;
